[FEAT] 상품 목록 응답 형식 통합 (#5)
