--- app/src/main/java/com/refit/app/data/product/model/Product.kt
@@ -1,11 +1,0 @@
-package com.refit.app.data.product.model
-
-data class Product(
-    val id: Long,
-    val image: String,
-    val brand: String,
-    val name: String,
-    val discountRate: Int,
-    val price: Int,
-    val discountedPrice: Int
-)
