--- app/src/main/java/com/refit/app/data/product/model/ProductUiState.kt
@@ -1,10 +1,0 @@
-package com.refit.app.data.product.model
-
-data class ProductUiState(
-    val items: List<Product> = emptyList(),
-    val totalCount: Int = 0,
-    val hasMore: Boolean = false,
-    val nextCursor: String? = null,
-    val isLoading: Boolean = false,
-    val error: String? = null
-)
