--- app/src/main/java/com/refit/app/data/product/model/ProductsPage.kt
@@ -1,8 +1,0 @@
-package com.refit.app.data.product.model
-
-data class ProductsPage(
-    val items: List<Product>,
-    val totalCount: Int,
-    val hasMore: Boolean,
-    val nextCursor: String?
-)
